Fixes imports and moves common tests from workspace_test.js to a helper file.
