+++ tests/mocha/generator_test.js
@@ -6,14 +6,13 @@
 
 goog.module('Blockly.test.generator');
 
-const {sharedTestSetup, sharedTestTeardown} = goog.require('Blockly.test.helpers');
-
-
 goog.require('Blockly.Dart');
 goog.require('Blockly.JavaScript');
 goog.require('Blockly.Lua');
 goog.require('Blockly.PHP');
 goog.require('Blockly.Python');
+const {sharedTestSetup, sharedTestTeardown} = goog.require('Blockly.test.helpers');
+
 
 suite('Generator', function() {
   setup(function() {
